DEV: Make TestPlugin common fixture
